Moved unused storage traits to unit tests
