+++ tests/statement_serializer_tests/foreign_key.cpp
@@ -3,6 +3,8 @@
 
 #include <type_traits>  //  std::is_same
 
+#include "../static_tests/static_tests_storage_traits.h"
+
 using namespace sqlite_orm;
 
 #if SQLITE_VERSION_NUMBER >= 3006019
@@ -330,6 +332,8 @@ TEST_CASE("statement_serializer foreign key") {
         REQUIRE(value == R"(FOREIGN KEY("user_id") REFERENCES "users"("id"))");
     }
     SECTION("composite key") {
+        using namespace sqlite_orm::internal::storage_traits;
+
         struct User {
             int id = 0;
             std::string firstName;
@@ -354,17 +358,16 @@ TEST_CASE("statement_serializer foreign key") {
                                      make_column("last_name", &User::lastName),
                                      primary_key(&User::id, &User::firstName));
 
-        STATIC_REQUIRE(internal::storage_traits::table_foreign_keys_count<decltype(usersTable), User>::value == 0);
-        STATIC_REQUIRE(internal::storage_traits::table_foreign_keys_count<decltype(usersTable), UserVisit>::value == 0);
+        STATIC_REQUIRE(table_foreign_keys_count<decltype(usersTable), User>::value == 0);
+        STATIC_REQUIRE(table_foreign_keys_count<decltype(usersTable), UserVisit>::value == 0);
 
         auto visitsTable = make_table("visits",
                                       make_column("user_id", &UserVisit::userId),
                                       make_column("user_first_name", &UserVisit::userFirstName),
                                       make_column("time", &UserVisit::time),
                                       fk);
-        STATIC_REQUIRE(internal::storage_traits::table_foreign_keys_count<decltype(visitsTable), User>::value == 1);
-        STATIC_REQUIRE(internal::storage_traits::table_foreign_keys_count<decltype(visitsTable), UserVisit>::value ==
-                       0);
+        STATIC_REQUIRE(table_foreign_keys_count<decltype(visitsTable), User>::value == 1);
+        STATIC_REQUIRE(table_foreign_keys_count<decltype(visitsTable), UserVisit>::value == 0);
 
         using storage_impl_t = internal::storage_impl<decltype(usersTable), decltype(visitsTable)>;
