+++ tests/static_tests/select_return_type.cpp
@@ -2,6 +2,7 @@
 #include <catch2/catch.hpp>
 
 #include "static_tests_common.h"
+#include "static_tests_storage_traits.h"
 
 using namespace sqlite_orm;
 
@@ -20,13 +21,13 @@ TEST_CASE("Select return types") {
         using IdsTuple = SelectVectorTuple::value_type;
         STATIC_REQUIRE(std::tuple_size<IdsTuple>::value == 1);
     }
+    //  test storage traits
     {
-        //  test storage traits
+        using namespace sqlite_orm::internal::storage_traits;
         struct Visit {
             int id = 0;
             std::string date;
         };
-        using namespace sqlite_orm::internal::storage_traits;
 
         //  test is_mapped
         STATIC_REQUIRE(is_mapped_v<decltype(storage), User>);
